// types/deviantart.d.ts
type TODO = any;
// Generated by https://quicktype.io

export interface SearchPopularResults {
  has_more: boolean;
  next_offset: number;
  results: Result[];
}

export interface Result {
  deviationid: string;
  printid: null | string;
  url: string;
  title: string;
  category: string;
  category_path: string;
  is_downloadable: boolean;
  is_mature: boolean;
  is_favourited: boolean;
  is_deleted: boolean;
  author: Author;
  stats: Stats;
  published_time: number;
  allows_comments: boolean;
  preview: Content;
  content: Content;
  thumbs: Content[];
}

export interface Author {
  userid: string;
  username: string;
  usericon: string;
  type: string;
}

export interface Content {
  src: string;
  filesize?: number;
  height: number;
  width: number;
  transparency: boolean;
}

export interface Stats {
  comments: number;
  favourites: number;
}