// types/google.d.ts
// Generated by https://quicktype.io

export interface GoogleCustomSearchResponse {
  kind: string;
  url: URL;
  queries: Queries;
  context: Context;
  searchInformation: SearchInformation;
  items: Item[];
}

export interface Context {
  title: string;
}

// Generated by https://quicktype.io

export interface Item {
  kind: string;
  title: string;
  htmlTitle: string;
  link: string;
  displayLink: string;
  snippet: string;
  htmlSnippet: string;
  mime: string;
  fileFormat: string;
  image: Image;
}

export interface Image {
  contextLink: string;
  height: number;
  width: number;
  byteSize: number;
  thumbnailLink: string;
  thumbnailHeight: number;
  thumbnailWidth: number;
}

export enum DisplayLink {
  WWWDeviantartCOM = "www.deviantart.com",
  WWWInstagramCOM = "www.instagram.com",
  WWWPinterestCOM = "www.pinterest.com",
}

export enum Kind {
  CustomsearchResult = "customsearch#result",
}

export interface Pagemap {
  cse_thumbnail?: CSEThumbnail[];
  metatags: { [key: string]: string }[];
  cse_image: CSEImage[];
  xfn?: Xfn[];
  ImageObject?: ImageObject[];
}

export interface ImageObject {
  acquireLicensePage: string;
  contentUrl: string;
  copyrightNotice: string;
  creditText: string;
}

export interface CSEImage {
  src: string;
}

export interface CSEThumbnail {
  src: string;
  width: string;
  height: string;
}

export interface Xfn {}

export interface Queries {
  request: NextPage[];
  nextPage: NextPage[];
}

export interface NextPage {
  title: string;
  totalResults: string;
  searchTerms: string;
  count: number;
  startIndex: number;
  inputEncoding: string;
  outputEncoding: string;
  safe: string;
  cx: string;
}

export interface SearchInformation {
  searchTime: number;
  formattedSearchTime: string;
  totalResults: string;
  formattedTotalResults: string;
}

export interface URL {
  type: string;
  template: string;
}